pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-tanishi"
version = "0.1.0"
description = "Autonomous Personal AI System — Sarcastic, Self-Improving, Loyal"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Private"}

dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "aiosqlite>=0.20.0",
    "cryptography>=43.0.0",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
memory = ["chromadb>=0.5.0", "mem0ai>=0.1.0"]
voice = ["faster-whisper>=1.0.0", "pipecat-ai>=0.5.0"]
all = ["project-tanishi[memory,voice]"]

[project.scripts]
tanishi = "tanishi.cli:main"
tanishi-server = "tanishi.api.server:main"

[tool.setuptools.packages.find]
include = ["tanishi*"]
